thor: Remove Digital Metadata that duplicated Digital RF metadata.
We want there to be one canonical place to find any piece of metadata
(e.g. samples_per_second) and thankfully nothing has come to rely on the
duplicated metadata yet.

This also adds the UUID string for the experiment to the Digital
Metadata, as promised in the documentation as the purpose of the UUID.