Unify glob/file listing, add key to file sort for get_bounds.
This fixes a bug with DigitalRFReader and DigitalMetadataReader
get_bounds where the bounds were incorrect if the number of digits in
the sample timestamp in the file name changed. This happened because the
sort was sorting by name and not by the value of the timestamp, which is
what we actually want (they are the same for the same number of digits).