Make reading metadata more convenient.
Adds a `method` argument to the read method of DigitalMetadataReader
to allow forward filling of values. Adds a `read_metadata` method to
DigitalRFReader to provide a convenient way to get the customary Digital
Metadata that accompanies an RF channel.

Also fix up the examples and tools to use these features and be more
uniform in how they read metadata.