Digital RF writer: Improvements in preparation for new GNU Radio sink.
Enable writing gapped data through the python interface even when the
recording mode is continuous. Add UUID generation if unspecified
to the python writer. Clarify docstrings and comments to make it clear
that 'is_continuous' is a data storage mode and not a requirement on
the data to be written. Also touch up docstrings/comments in general.
Add a case to the python benchmark and test scripts to looked at chunked
writes and gapped input data, respectively.