digital_rf_sink: Simplify some of the logic and hopefully speed it up.
Now if `start` is None, we default to a start sample index of 0 instead
of ignoring data until a time tag arrives. This simplifies the code and
should be more intuitive for the user.

This also simplifies the stream tag parsing by reading rx_freq and
metadata tags in chunks corresponding to the data blocks as marked by
rx_time tags. Also move to using lists to represent the data blocks
instead of an OrderedDict.