python: Rewrite tests to use pytest, increase comprehensiveness.
We now test many different values of DigitalRFWriter parameters and
exhaustive combinations of those parameters. We check object attributes
and methods for DigitalRFWriter and DigitalRFReader and exact
round-tripping of data through the write and read methods. In the
process of writing this test suite, we found a number of bugs involving
edge cases that were fixed in recent previous commits.