python: Leave pkgconfig out of pyproject.toml until pip can use source.
In its PEP 518 support, pip as of version 10.0.0b2 requires that all
build requirements be available as wheels so that build recursion
doesn't happen. Eventually they plan to do away with this requirement.
Until then (or until pkgconfig is available as a wheel), leave pkgconfig
out of the pyproject.toml to fix build failures with pip.