thor: Add [clock|time]_source options, wait_for_lock option.
This deprecates the global sync_source option in favor of per-mainboard
clock and time source options (since devices like the B200-mini can make
use of having separate settings for sources). Adding a wait_for_lock
option that allows skipping the ref_locked sensor check is also useful
for the B200-mini which has the sensor but might not have a reference to
lock to even if PPS syncing is desired.