Make libdigital_rf (data format) version track package version.
Instead of only incrementing the data format version to the next value
when the format changes, increment it so it syncs up with the package
version at the time of the change. This should overall be less confusing
as the format version will indicate the package version that it debuted
with.