thor: Start flowgraph ASAP, exit if no samples after launch.
With the increased sleep times after the dummy start/stop to create the
RX streamer object, it doesn't seem necessary to wait to start the
flowgraph. In fact, starting it ASAP helps prevent the error where it
starts too late and the intended launch time is already in the past.