Always store 2-D arrays in HDF5 files, read returns 2-D if no subchan.
We used to store arrays as 2-D if they were complex, and 1-D or 2-D
depending on the number of subchannels if they were real. For part of
this development cycle, I had switched to storing as 1-D if there was
only a single subchannel. Since this would break old reader software on
new data for the most common case of single subchannel complex, we're
going back to 2-D but also doing that in the real case for consistency.

The plain read method will now always return a 2-D array if no
subchannel is specified. All other read methods squeeze single
subchannel data down to 1-D when it is read for ease of use.