digital_metadata: Disable file locking for HDF5 >= 1.10.
HDF5 1.8 and earlier doesn't have file locking, so this restores
behavior of HDF5 >= 1.10 to match previous versions. Disabling file
locking is useful because we sometimes might want to be able to read and
write metadata using the same file. Even though we only open a file for
reading/writing when it is actively being read from or written to, we
want concurrent access to succeed on the rare occasions when
reading/writing happens at the same time. This should be safe because
we've been doing it in practice for a while now with HDF5 1.8 and
earlier. Additionally, we don't allow overwriting or adding to a sample
index, so all cases of concurrent read/write should be accessing
different groups in the file.