python/digital_rf_hdf5: Fix __del__ on partially __init__ed obj.
In _top_level_dir_properties, __del__ relies on the _cachedFile
attribute being set, but since it was set at the end of __init__, an
error earlier in __init__ that was caught could lead to __del__ being
called without _cachedFile being set. This moves the setting of
_cachedFile up before anything that could reasonably raise an error, so
it should always be set before __del__ is called.