python: Add pkgconfig to build dependencies in pyproject.toml.
This was left out earlier because pkgconfig was not available as a
wheel, and so pip was failing. Now that pkgconfig is a wheel (since
1.4.0), we can include this.

Also, this might fix the build error currently being seen with Travis
where HDF5 is not found when installing the Python package tarball,
presumably because pkgconfig is not installed and it therefore doesn't
know the path to HDF5.