thor: Fix resampling and channelizing filters.
The resampling filter previously used GNU Radio's Rational Resampler,
which as of now introduces an advance of the sample timing that is
unexpected and out of line with all of the other filter blocks. This
switches to using the Polyphase Arbitrary Resampler, which does nearly
the same thing (with perhaps some added overhead) but introduces filter
delays as expected. Incidentally, the delay that was applied was wrong
anyway, so this also fixes that.

The other enhancement here is to specify the filter delays for the
used filter blocks so that sample tags are propagated to account for the
delay. This ensures that the USRP time tags that are used to place the
output samples within the Digital RF file are correct.