digital_metadata: Return python types for scalars when reading.
This is helpful mostly because the numpy scalars should be compatible
with the corresponding python type, but they sometimes aren't (such as
an np.int_ not being a subclass of int on python 3). There's also a good
chance that the value was written from a python type in the first place,
so this would complete that round trip.

This also makes a departure from h5py by always returning ascii strings
as a python string type (str or unicode) and not as bytes. The data was
a string when it was written, so it is unintuitive that it becomes bytes
when read. Hopefully h5py changes their default behavior and recognizes
ascii as a string and not arbitrary bytes, but we need this until then.