digital_rf_sink: Fixes/improvements for stop_on_skipped == True.
Previously, stop_on_skipped would not work if the sample skip happened
to occur right at a work() boundary so that only one block of data was
being written but that the block was discontinuous from the last write.

Fixing that means that a "skip" at the very first write (e.g. when a
USRP time tag on the first sample is later than the requested start
time) would trigger the stopping, which is not desired behavior
considering that it's not really a skip but rather a delay from when the
recording was thought to start. This change also fixes that case by
ignoring the "skip" for the first write.