python: Add option to force polling for watchdog functions.
The default watchdog observer might not always work (perhaps depending
on the filesystem), sometimes failing silently and causing ringbuffer
and mirror to appear to do nothing. This adds an option to the affected
classes and command line tools to force polling the filesystem for
changes, which should always work as a backup option.