gr_digital_rf/digital_rf_sink: Fix possible IndexError with stop_on_*
When using the `stop_on_skipped` or `stop_on_time_tag` flags with
`gr_digital_rf.digital_rf_channel_sink`, an IndexError could occur by
trying to reference a second data block when there is only one data
block to be written before exiting after stopping is flagged. It was
possible to rewrite the offending code to not rely on indexing into the
list of data blocks.