watchdog_drf: fast path for _set_root, override stop, improve drf watch.
The improvement to 'drf watch' is to periodically check that the
observer is alive and create a new one if necessary. This catches the
case where the watched path is deleted; only some of the
watchdog emitters (only polling?) emit an event for that, and otherwise
the root_observer thread is stopped. This will catch the stopped thread
and recover, meaning that the path is still watched for re-creation even
after it has been deleted.