python/tests: Tweak writer test to fail with 64-bit long double.
If digital_rf is compiled to use 64-bit long double values (as opposed
to 80-bit or more), or other compiler flags throw off the long double
sample rate math, users are likely to see write failures depending on
the sample rate. Usually this results in an error message of
"Failed to write data" and "Request index M before first expected
index N".

The current Python tests don't trigger this condition, so they pass when
using 64-bit long doubles. This commit tweaks the file cadence so that
the write error is triggered in these cases, so we can be forewarned
that errors are likely when writing under those compilation conditions.