c, python: Expose integer math to calculate file sample bounds.
This does away with use of `np.longdouble`s for calculations with the
sample rate in the Python library, which was used to get the list of
files containing a given sample index range. See 90cdb49 for the commit
that does the same thing in the C library.

This commit also exposes two new functions to the C API:
`digital_rf_get_timestamp_floor` and `digital_rf_get_sample_ceil`. These
are now used to do the file sample bound calculations, and they are
exposed so that users of the library can perform these calculations in a
way that is consistent with what is done internally. These functions
have also been exposed in the Python binding, and a test that verifies
that the sample index round-trips through them is included in the Python
testing suite.

Because of the added C functions, the library version has been bumped to
2.7, which means the next release of Digital RF will be 2.7.0.