gr_digital_rf: Remove check for GNU Radio version.
This check fails when GNU Radio is built from a non-version-tagged git
commit. In that case, `gr.version()` returns a string that is different
from the tagged case and is not parseable by
`distutils.version.LooseVersion`, so the check fails with a type error.

Since we're not currently even doing the frequency tag parsing, this
check doesn't actually change anything. Also, the old GNU Radio versions
for which this would have mattered are now old enough that they are
unlikely to be seen in the wild. Thus, the simplest fix is to remove
this check on the GNU Radio version.