python: Change build dependency to oldest-supported-numpy.
Pip now defaults to isolated build mode, so the best way to build
against an old numpy version (for compatibility with that through newer
versions) is to specify it in pyproject.toml. Thankfully there is the
'oldest-supported-numpy' package that can use the oldest version for
whatever platform so we don't have to do it manually.