python: tools: Fix thor.py multi-channel recording with GR 3.9+.
GNU Radio versions starting with 3.9 (with pybind11 Python bindings
instead of SWIG) apparently no longer store object references to
blocks when they are connected to a flowgraph. This was causing failures
with thor.py when recording with multiple channels (`AttributeError:
'list_iterator' object has no attribute 'start'`) because the block
objects were built in a loop that did not keep its own object references
from previous iterations, which caused the blocks to be
garbage-collected before/during execution.

The fix here is to add all of the block objects to a list when
constructing the flowgraph in thor.py, so they will no longer be garbage
collected.