python: Fix _get_file_list for compat with numpy 1.23 on Windows.
Because of changes in how np.longdouble is handled on Windows (it has
always only had double precision, but now they're being more strict
about that somehow), it errors on an operation between a Python int and
a np.longdouble. This adds an explicit conversion from int to np.uint64
first.