digital_metadata: When reading, convert numpy.object_ array to list
This improves interoperability with GNU Radio pmt types which can
handle lists but not numpy object arrays. Additionally, an object
arrays coming from an HDF5 file stands a decent chance of having
originally been a list when it was passed in for writing, so this allows
a round trip to the same type.

In the case that the object array holds strings, convert them from bytes
to str type in much the same way we do for scalar values.