python/watchdog_drf: Fix checking for dest_path in FileSystemEvent
Prior to watchdog 4.0.0, FileSystemEvent only had a dest_path attribute
if it was a move event. Since 4.0.0, all FileSystemEvents have a
dest_path attribute, defaulting to an empty string when not a move
event.

Because we were checking for dest_path is None, with watchdog 4.0.0+
non-move events were triggering the check for dest_path and getting
changed to a FileDeletedEvent. This fixes that check so that it works
for all versions of watchdog.