Upload tarball to github from wheel building workflow
The new release_tarball workflow did not work quite right, in that the
tarball lacked package information (including the version) that broke
the build from source (when not building from a git repo). But that
means we just need the same source tarball that gets uploaded to PyPI,
so just push that to GitHub as well.